+++ lib/mongoid/orderable/handlers/document.rb
@@ -6,17 +6,10 @@ module Handlers
   class Document < Base
 
     def before_create
-      if use_transactions
-        clear_all_positions
-      else
-        apply_all_positions
-      end
-    end
-
-    def after_create
-      return unless use_transactions
       apply_all_positions
     end
+
+    def after_create; end
 
     def before_update
       return unless any_field_changed?
@@ -25,12 +18,6 @@ module Handlers
 
     def after_destroy
       remove_all_positions
-    end
-
-    protected
-
-    def clear_all_positions
-      orderable_keys.each {|field| doc.send("orderable_#{field}_position=", nil) }
     end
   end
 end
