+++ lib/mongoid/orderable/handlers/document_embedded.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Mongoid
+module Orderable
+module Handlers
+  class DocumentEmbedded < Document
+
+    def after_destroy
+      return if doc._root.destroyed?
+      super
+    end
+  end
+end
+end
+end
