+++ lib/mongoid_orderable.rb
@@ -27,6 +27,7 @@ require 'mongoid/orderable/generators/scope'
 require 'mongoid/orderable/generators/helpers'
 require 'mongoid/orderable/handlers/base'
 require 'mongoid/orderable/handlers/document'
-require 'mongoid/orderable/handlers/embedded'
+require 'mongoid/orderable/handlers/document_embedded'
+require 'mongoid/orderable/handlers/document_transactional'
 require 'mongoid/orderable/handlers/transaction'
 require 'mongoid/orderable/installer'
